ref: Move events specific processor code into EventsProcessor
Closes #336
