+++ tests/test_consumer.py
@@ -146,8 +146,8 @@ class TestConsumer(BaseTest):
         test_worker = ConsumerWorker(self.clickhouse, self.dataset, producer, self.dataset.get_default_replacement_topic())
 
         test_worker.flush_batch([
-            (processor.REPLACE, ('1', {'project_id': 1})),
-            (processor.REPLACE, ('2', {'project_id': 2})),
+            (self.dataset.get_processor().REPLACE, ('1', {'project_id': 1})),
+            (self.dataset.get_processor().REPLACE, ('2', {'project_id': 2})),
         ])
 
         assert [(m._topic, m._key, m._value) for m in producer.messages] == \
