+++ tests/test_writer.py
@@ -1,7 +1,6 @@
 from base import BaseEventsTest
 
 from snuba.clickhouse import ColumnSet
-from snuba.processor import process_message
 
 
 class TestWriter(BaseEventsTest):
@@ -13,7 +12,7 @@ class TestWriter(BaseEventsTest):
         assert res[0][0] == 1
 
     def test_columns_match_schema(self):
-        _, processed = process_message(self.dataset.get_promoted_tag_columns(), self.event)
+        _, processed = self.dataset.get_processor().process_message(self.event)
         row = self.dataset.row_from_processed_message(processed)
 
         # verify that the 'count of columns from event' + 'count of columns from metadata'
@@ -25,7 +24,7 @@ class TestWriter(BaseEventsTest):
         """Fields in a processed events are ignored if they don't have
         a corresponding Clickhouse column declared."""
 
-        _, processed = process_message(self.dataset.get_promoted_tag_columns(), self.event)
+        _, processed = self.dataset.get_processor().process_message(self.event)
 
         processed['unknown_field'] = "unknown_value"
         row = self.dataset.row_from_processed_message(processed)
